project_files/smart_energy_consumption_optimizer_20260402_174938/smart_energy_consumption_optimizer/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pandas==2.0.3
numpy==1.24.3
prophet==1.1.1
streamlit==1.26.0
python-dotenv==0.21.1
pydantic==2.3.1
matplotlib==3.7.1
plotly==5.18.0
gunicorn==20.1.0
python-multipart==0.0.5
typing-extensions==4.5.0